update merk tests
