+++ merk/src/proofs/chunk.rs
@@ -583,6 +583,8 @@ mod tests {
             .unwrap()
             .unwrap();
 
+        merk.commit();
+
         let root_node = merk.tree.take();
         let root_key = root_node.as_ref().unwrap().key().to_vec();
         merk.tree.set(root_node);
